CarrotSquare bug fixed. issue with no valid move sorted (player can now stay on carrot square when they don't have enough carrots to move forward and tortoise square is occupied) and player is not sent back to the start.
Code for this in GameController is awful, but it works. If you want to have a look and change it, be my guest.

Had to make changes to CarrotSquare Class in order to remove option 3 for 'player to take turn as normal' when they have no carrots. otherwise the game would be stuck as player couldn't make a move. Option 3 now only available when player can make a valid move.